feat: Add multi-phase pipeline with modularized training
- Modularize __main__.py into training/ package (constants, metrics,
  checkpoint, memory, models, agents, reward, progress, runner)
- Add Phase 0a (APK manifest parsing via aapt2)
- Add Phase 0b (semantic crawling with LLM annotation + auto
  requirements.csv with XML cross-validation)
- Add Phase 0c (replay memory warmup from crawl transitions)
- Add Phase 2 (enriched transcription with screenshots)
- Add metrics/ package (CoverageTracker, PhaseObserver, SBSE bridge)
- Add report/ package (HTML reporter with SVG gauges)
- Add functional reward (element coverage + input class bonuses)
- Add CrewAI prompt rules to avoid raw resource-ids in test cases
- Add train-phases CLI command
- Fix Phase 0c tensor device mismatch (LongTensor instead of torch.tensor)
- Fix Phase 0c action index out of bounds (cap at agent.num_actions)

+++ rlmobtest/models/dqn_model.py
@@ -86,9 +86,7 @@ def select_action(state, actions):
     """Select an action using epsilon-greedy policy."""
     global steps_done
     sample = random.random()
-    eps_threshold = EPS_END + (EPS_START - EPS_END) * math.exp(
-        -1.0 * steps_done / EPS_DECAY
-    )
+    eps_threshold = EPS_END + (EPS_START - EPS_END) * math.exp(-1.0 * steps_done / EPS_DECAY)
     steps_done += 1
     if sample > eps_threshold:
         with torch.no_grad():
@@ -115,9 +113,9 @@ def optimize_model():
     batch = Transition(*zip(*transitions))
 
     non_final_mask = BoolTensor(tuple(map(lambda s: s is not None, batch.next_state)))
-    non_final_next_states_t = torch.cat(
-        tuple(s for s in batch.next_state if s is not None)
-    ).type(dtype)
+    non_final_next_states_t = torch.cat(tuple(s for s in batch.next_state if s is not None)).type(
+        dtype
+    )
 
     with torch.no_grad():
         non_final_next_states = Variable(non_final_next_states_t)
